kds: introduce TCBVersionV0 for StructVersion 0
AMD Specification #57230 defines distinct 8-byte hardware wire format layouts
(`structVersion 0` for Family 19h vs `structVersion 1` for Family 1Ah / Turin).
The legacy monolithic `TCBParts` struct cannot represent both hardware schemas.

Replace `TCBParts` with an explicit `TCBVersionV0` type for StructVersion 0
(Milan, Genoa, Siena) as a preparatory refactor before introducing Turin support.

+++ testing/fakekds.go
@@ -187,7 +187,7 @@ func (f *FakeKDS) Get(url string) ([]byte, error) {
 	if certs == nil {
 		return nil, fmt.Errorf("no certificate found at %q (unknown HWID %v)", url, vcek.HWID)
 	}
-	certbytes, ok := certs[vcek.TCB]
+	certbytes, ok := certs[uint64(vcek.TCB)]
 	if !ok {
 		return nil, fmt.Errorf("no certificate found at %q (host present, bad TCB %v)", url, vcek.TCB)
 	}
